fix: complete uncompleted work [refactore code structure of terminator cli & add gemini and openai support to terminato-cli chat]
