+++ terminator-cli/src/mpc_client/mod.rs
@@ -1,0 +1,6 @@
+pub mod utils;
+pub mod gemini;
+pub mod openai;
+pub mod anthropic;
+pub mod natural_lang;
+pub mod interactive_chat;
